refactor: split core domain services
