+++ src/core/index.ts
@@ -17,5 +17,8 @@ export * from "./models";
 export * from "./people";
 export * from "./ports";
 export * from "./queries";
-export * from "./tracker-services";
+export * from "./gig-domain-service";
+export * from "./task-domain-service";
+export * from "./artifact-domain-service";
+export * from "./deep-patch";
 export * from "./services";
